pyproject.toml
[tool.poetry]
name = "homehuntr"
version = "0.1.0"
description = ""
authors = ["jeffbrennan <jeffbrennan10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
polars = "^0.19.19"
pyspark = "^3.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
